backend/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asgiref==3.9.1
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
auth==0.5.3
bcrypt==4.0.1
bidict==0.23.1
blinker==1.9.0
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cryptography==46.0.1
cycler==0.12.1
Cython==3.1.4
cytoolz==1.0.1
dataclasses-json==0.6.7
datasets==4.1.1
decorator==5.2.1
deepgram-sdk==5.0.0
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
email-validator==2.3.0
eventlet==0.40.3
exceptiongroup==1.3.0
executing==2.2.1
falcon==4.1.0
fiddle==0.3.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
flask-sock==0.7.0
Flask-SocketIO==5.5.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2024.12.0
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-speech==2.33.0
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.75.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
indic_numtowords==1.1.0
inflect==7.5.0
intervaltree==3.1.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jiwer==3.1.0
joblib==1.5.2
julius==0.2.7
kaldi-python-io==1.2.2
kiwisolver==1.4.9
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.31.1
libcst==1.8.5
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.15.2
lilcom==1.8.1
llvmlite==0.45.1
loguru==0.7.3
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
ml_dtypes==0.5.3
mongoengine==0.29.1
more-itertools==10.8.0
motor==3.7.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nemo-toolkit==2.5.0
networkx==3.4.2
num2words==0.5.14
numba==0.62.1
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.19.0
openai==1.109.1
optuna==4.5.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
peft==0.17.1
pexpect==4.9.0
pillow==11.3.0
plac==1.4.5
platformdirs==4.4.0
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyloudnorm==0.1.1
pymongo==4.15.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-http-client==3.3.7
python-socketio==5.13.0
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
redis==4.6.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.4.3
rich==14.1.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sendgrid==6.12.5
sentencepiece==0.2.1
sentry-sdk==2.40.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
speechbrain==1.0.3
SQLAlchemy==2.0.43
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.1.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
toolz==1.0.0
torch==2.2.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.2.2
torchmetrics==1.8.2
torchvision==0.17.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
typeguard==4.4.4
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wandb==0.22.1
wcwidth==0.2.14
webdataset==1.0.2
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
whisper_normalizer==0.1.12
wrapt==1.17.3
wsproto==1.2.0
xxhash==3.6.0
yarl==1.20.1
zope.event==6.0
zope.interface==8.0.1
